build: game score data type Double -> BigDecimal

+++ src/main/kotlin/kr/jiasoft/hiteen/feature/play/dto/GameScoreWithParticipantView.kt
@@ -1,5 +1,6 @@
 package kr.jiasoft.hiteen.feature.play.dto
 
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -8,7 +9,7 @@ data class GameScoreWithParticipantView(
     val participantId: Long,
     val userId: Long,
     val league: String,
-    val score: Double,
+    val score: BigDecimal,
     val tryCount: Int?,
     val createdAt: OffsetDateTime,
     val userNickname: String,
